+++ internal/cli/enterprise.go
@@ -19,6 +19,7 @@ import (
 	"github.com/urfave/cli/v2"
 
 	"github.com/redpanda-data/connect/v4/internal/impl/kafka/enterprise"
+	"github.com/redpanda-data/connect/v4/internal/license"
 	"github.com/redpanda-data/connect/v4/internal/secrets"
 	"github.com/redpanda-data/connect/v4/internal/telemetry"
 )
@@ -46,6 +47,9 @@ func InitEnterpriseCLI(binaryName, version, dateBuilt string, schema *service.Co
 	}
 
 	var disableTelemetry bool
+	licenseConfig := license.Config{
+		LicenseFilepath: os.Getenv("REDPANDA_LICENSE_FILEPATH"),
+	}
 
 	opts = append(opts,
 		service.CLIOptSetVersion(version, dateBuilt),
@@ -81,6 +85,9 @@ func InitEnterpriseCLI(binaryName, version, dateBuilt string, schema *service.Co
 		}),
 		service.CLIOptAddTeeLogger(slog.New(rpLogger)),
 		service.CLIOptOnConfigParse(func(pConf *service.ParsedConfig) error {
+			// Kick off license service.
+			license.RegisterService(pConf.Resources(), licenseConfig)
+
 			// Kick off telemetry exporter.
 			if !disableTelemetry {
 				telemetry.ActivateExporter(instanceID, version, fbLogger, schema, pConf)
@@ -103,8 +110,13 @@ func InitEnterpriseCLI(binaryName, version, dateBuilt string, schema *service.Co
 				Name:  "disable-telemetry",
 				Usage: "Disable anonymous telemetry from being emitted by this Connect instance.",
 			},
+			&cli.StringFlag{
+				Name:  "redpanda-license",
+				Usage: "Provide an explicit Redpanda License, which enables enterprise functionality. By default licenses found at the path `/etc/redpanda/redpanda.license` are applied.",
+			},
 		}, func(c *cli.Context) error {
 			disableTelemetry = c.Bool("disable-telemetry")
+			licenseConfig.License = c.String("redpanda-license")
 
 			if secretsURNs := c.StringSlice("secrets"); len(secretsURNs) > 0 {
 				var err error