+++ internal/impl/openai/transcription_processor.go
@@ -16,6 +16,8 @@ import (
 	"github.com/redpanda-data/benthos/v4/public/bloblang"
 	"github.com/redpanda-data/benthos/v4/public/service"
 	oai "github.com/sashabaranov/go-openai"
+
+	"github.com/redpanda-data/connect/v4/internal/license"
 )
 
 const (
@@ -65,6 +67,10 @@ To learn more about audio transcription, see the: https://platform.openai.com/do
 }
 
 func makeTranscriptionProcessor(conf *service.ParsedConfig, mgr *service.Resources) (service.Processor, error) {
+	if err := license.CheckRunningEnterprise(mgr); err != nil {
+		return nil, err
+	}
+
 	b, err := newBaseProcessor(conf)
 	if err != nil {
 		return nil, err